Add explicit casts and use CKUINTPTR consistently
Update CKDataArray API to use CKUINTPTR for key parameters instead of
CKDWORD to properly handle pointer-sized values on 64-bit platforms.
Add explicit casts for narrowing conversions from size_t to int to
address compiler warnings.

Also fixes XObjectArray::RemoveObject to correctly calculate the index
of the removed object using pointer arithmetic.

+++ src/CKParameterManager.cpp
@@ -647,7 +647,7 @@ CKERROR CKParameterManager::RegisterNewStructure(CKGUID StructGuid, CKSTRING Str
     const char *currentPos = StructData;
     for (int i = 0; i < memberCount; ++i) {
         const char *end = strchr(currentPos, ',');
-        int len = end ? (end - currentPos) : (int) strlen(currentPos);
+        int len = end ? (int)(end - currentPos) : (int) strlen(currentPos);
         structStruct.Desc[i] = new char[len + 1];
         strncpy(structStruct.Desc[i], currentPos, len);
         structStruct.Desc[i][len] = '\0';
@@ -722,7 +722,7 @@ CKERROR CKParameterManager::RegisterNewStructure(CKGUID StructGuid, CKSTRING Str
     const char *currentPos = StructData;
     for (int i = 0; i < memberCount; ++i) {
         const char *end = strchr(currentPos, ',');
-        int len = end ? (end - currentPos) : (int) strlen(currentPos);
+        int len = end ? (int)(end - currentPos) : (int) strlen(currentPos);
         structStruct.Desc[i] = new char[len + 1];
         strncpy(structStruct.Desc[i], currentPos, len);
         structStruct.Desc[i][len] = '\0';
